Improve names and version numbers for Modrinth versions
This improves the names and version numbers for versions on the Modrinth project.
First, it changes the version name from `Inventory Profiles Next-$mod_loader-$minecraft_version-$mod_version` to `IPN $mod_version for $mod_loader $minecraft_version`; this means, instead of being displayed as `Inventory Profiles Next-fabric-1.17-0.8.3`, it will be displayed as `IPN 0.8.3 for fabric 1.17`.
It also changes the version number listed to match the name of the jar. Instead of the version number being `Inventory Profiles Next-$mod_loader-$minecraft_version-$mod_version`, it will now be `$mod_loader-$minecraft_version-$mod_version`; this means, instead of being displayed as `Inventory Profiles Next-fabric-1.17-0.8.3`, the version number will be displayed as `fabric-1.17-0.8.3`.
Why is this important? Firstly, it improves readability for the end user and for anyone looking at the mod's versions.
Secondly, it allows mod developers to use Modrinth's Maven functionality with InvProNext. Because Gradle is very weird, it does not allow the use of artifacts with spaces in the name (URI encoded or not). By removing the spaces and removing the overall strangeness of the version number, it will allow developers to leverage Modrinth's Maven in order to include it in their development environments.
In case you want to document the use of Modrinth's Maven, the repository is located at `https://api.modrinth.com/maven`, and the artifact (after this PR) would be `maven.modrinth:inventory-profiles-next:fabric-1.17-0.8.3`.
By the way - thank you for maintaining InvPro for these versions, and thank you for using Modrinth! :smile: This is the only thing that's kind of irked me; otherwise, thank you greatly and keep up the great work. :slight_smile: